Changed the autospeed -1 by dreambox
For dreambox now It will be in autospeed if autospeed = 1 (or ommited in configs default value)
In web if autospeed box checked.
Addapted the comments.
The 2700 3150 and 8300 remains as there is no other way to differ the stb.
To set a manual clockspeed just uncheck the autospeed box first. or set the autospeed = 0 in configs